Lifted out base class to its own file
